+++ projects/ngx-location-picker/src/lib/components/ngx-location-picker.component.ts
@@ -9,6 +9,9 @@ import {CoordinateModel} from '../types/coordinate.model';
 import {NotificationModel} from '../types/notification.model';
 import {NgxLocationPickerHelper} from '../services/ngx-location-picker.helper';
 import {LeafletTileLayerModel, LeafletTileLayerType} from '../types/leaflet-tile-layer.model';
+import {debounceTime, distinctUntilChanged} from 'rxjs/operators';
+import {Subject} from 'rxjs';
+import {InitialLocationModel} from '../types/initial-location.model';
 
 @Component({
   selector: 'aui-location-picker',
@@ -87,12 +90,17 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
   @Input() locationsLimit = 5;
   /* The layers to search locations for */
   @Input() locationLayers = ['straatnaam'];
-  /* Prioritize a layer, boosts results from a given layer to the top of the found locations. */
-  @Input() prioritizeLayer = 'straatnaam';
-  /* Sort locations by certain key, overrides prioritizeLayer. */
+  /**
+   * Prioritize specific layers, boosts results from given layers to the top of the found locations.
+   * The order of the values in the array determines the priority. Overrides sortBy.
+   */
+  @Input() prioritizeLayers = ['straatnaam'];
+  /* Sort locations by certain key. */
   @Input() sortBy = '';
   /* Use geolocation when the component finished loading */
   @Input() locateUserOnInit = false;
+  /* Set time to wait after user stops typing before triggering a search */
+  @Input() debounceTime = 200;
   /* AddPolygon event */
   @Output() addPolygon = new EventEmitter<any>();
   /* AddLine event */
@@ -123,6 +131,8 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
   /* Current tile layer type default or custom */
   tileLayerType: LeafletTileLayerType = LeafletTileLayerType.DEFAULT;
 
+  /* Input change subject */
+  private searchQueryChanged: Subject<string> = new Subject<string>();
   /* Current leaflet state */
   private mapLoaded = false;
   /* Keep a reference to our geolocation object */
@@ -147,16 +157,19 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
   private activeTileLayers = [];
 
   /* Used for ControlValueAccessor */
-  propagateChange = (_: any) => {};
+  propagateChange = (_: any) => {
+  };
 
   get selectedLocation() {
     return this._selectedLocation;
   }
 
   set selectedLocation(location) {
-    if (this.mapLoaded && location && location.position && location.position.lat) {
+    if (this.mapLoaded && location && location.position && location.position.lat && location.position.lng) {
       this.writeValue({}, true);
       this.setInitialLocation(location);
+
+      delete location.options;
 
       location = {
         ...location,
@@ -204,7 +217,13 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
     private locationPickerHelper: NgxLocationPickerHelper,
     private mapService: MapService,
     private renderer: Renderer2,
-  ) {}
+  ) {
+    this.inputChangeSubscription = this.searchQueryChanged
+      .pipe(debounceTime(this.debounceTime), distinctUntilChanged())
+      .subscribe((query) => {
+        this.onSearch(query);
+      });
+  }
 
   /**
    * On component init
@@ -279,6 +298,13 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
    */
   zoomOut() {
     this.leafletMap.zoomOut();
+  }
+
+  /**
+   * On input change event
+   */
+  onInputChange(query: string) {
+    this.searchQueryChanged.next(query);
   }
 
   /**
@@ -398,7 +424,7 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
         this.baseUrl,
         this.locationsLimit,
         this.locationLayers,
-        this.prioritizeLayer,
+        this.prioritizeLayers,
         this.sortBy
       ).subscribe((response: LocationModel[] | AddressModel[] | CoordinateModel[]) => {
         this.foundLocations = response;
@@ -603,13 +629,17 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
   /**
    * Triggers a search when selectedLocation was updated from outside our component.
    */
-  private setInitialLocation(initialLocation) {
+  private setInitialLocation(initialLocation: InitialLocationModel) {
     this.cancelGeolocation();
 
-    this.onSearch(
-      `${initialLocation.position.lat},${initialLocation.position.lng}`,
-      (initialLocation.label && this.locationPickerHelper.isCoordinate(initialLocation.label))
-    );
+    if (!('options' in initialLocation) || !('triggerSearch' in initialLocation.options) || initialLocation.options.triggerSearch) {
+      this.onSearch(
+        `${initialLocation.position.lat},${initialLocation.position.lng}`,
+        (initialLocation.label && this.locationPickerHelper.isCoordinate(initialLocation.label))
+      );
+    } else {
+      this.addMapMarker([initialLocation.position.lat, initialLocation.position.lng]);
+    }
   }
 
   /**
@@ -626,12 +656,13 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
     }
   }
 
+
   /**
    * Sets a dynamically fetched location when using locate-me or pick location on map
    */
   private setLocationDynamically(lat, lng) {
     this.resetFoundLocations();
-    this.onSearch(`${lat},${lng}`);
+    this.onSearch(`${lat.toFixed(6)},${lng.toFixed(6)}`);
   }
 
   /**
@@ -678,7 +709,7 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
 
     this.selectedLocationMarker.on('dragend', (event) => {
       const newCoords = this.selectedLocationMarker.getLatLng();
-      const searchValue = (newCoords) ? `${newCoords.lat},${newCoords.lng}` : '';
+      const searchValue = (newCoords) ? `${newCoords.lat.toFixed(6)},${newCoords.lng.toFixed(6)}` : '';
 
       this.pickedLocation = true;
       this.resetFoundLocations();
@@ -799,6 +830,6 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
 
     setTimeout(() => {
       this.leafletNotification = null;
-    }, 4000);
+    }, 3000);
   }
 }