+++ CHANGELOG.md
@@ -13,6 +13,14 @@ All notable changes to this project will be documented in this file.
 "### Security" in case of vulnerabilities.
 -->
 
+## [Unreleased]
+- [ADDED] debounce time for search queries
+- [CHANGED] notification position on leaflet
+- [ADDED] proper rounding on coordinate sets (lat/lng)
+- [ADDED] option to disable triggering a search when providing an initial location
+- [CHANGED] prioritizeLayer renamed to prioritizeLayers and now uses an array of layer names
+
+
 ## [4.3.1] - 2020-03-24
 
 ### Fixed